Update apps with @task() decorator (#121)
